Voice Rework

+++ core/src/synth/public/gen.rs
@@ -10,7 +10,12 @@ impl Synth {
     parameter numbers and ranges are described in the SoundFont 2.01
     specification, paragraph 8.1.3, page 48.
      */
-    pub fn set_gen(&mut self, chan: u8, param: GenParam, value: f32) -> Result<(), &'static str> {
+    pub fn set_gen(
+        &mut self,
+        chan: usize,
+        param: GenParam,
+        value: f32,
+    ) -> Result<(), &'static str> {
         if let Some(channel) = self.channels.get_mut(chan as usize) {
             channel.gen[param as usize] = value;
             channel.gen_abs[param as usize] = 0;
