+++ src/synth/gen.rs
@@ -15,7 +15,7 @@ impl Synth {
     parameter numbers and ranges are described in the SoundFont 2.01
     specification, paragraph 8.1.3, page 48.
      */
-    pub fn set_gen(&mut self, chan: u8, param: GenParam, value: f32) -> Result<(), &str> {
+    pub fn set_gen(&mut self, chan: usize, param: GenParam, value: f32) -> Result<(), &str> {
         self.handle.set_gen(chan, param, value)
     }
